requirements/financial.txt
# Financial App Dependencies
# Additional packages required for financial management functionality

# File Processing
pandas>=2.0.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
python-magic>=0.4.27

# Data Analysis and Visualization
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Financial Calculations
python-dateutil>=2.8.2
pytz>=2023.3

# HTTP and API
requests>=2.31.0
urllib3>=2.0.0

# Background Processing
celery>=5.3.0
redis>=4.5.0

# Caching and Performance
django-redis>=5.3.0
django-cacheops>=8.0.0

# Security and Authentication
django-ldap3>=2.9.0
django-radius>=1.0.0
django-oauth-toolkit>=1.6.0

# Reporting and Export
reportlab>=4.0.0
xlsxwriter>=3.1.0
openpyxl>=3.1.0

# Machine Learning for Categorization
scikit-learn>=1.3.0
nltk>=3.8.0

# Database Optimization
django-debug-toolbar>=4.1.0
django-extensions>=3.2.0

# Testing
factory-boy>=3.3.0
faker>=19.0.0
pytest-django>=4.5.0
pytest-cov>=4.1.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
isort>=5.12.0